Modularize meeting accessibility detection
Separate public data contracts, AX node semantics, and meeting app/platform classification from capture and send orchestration while preserving the detect API and behavior.
